{"type":"TI", "runs":"0"}

+++ Breakout.java
@@ -270,7 +270,7 @@ public class Breakout extends GraphicsProgram {
 	
 	public void mouseMoved(MouseEvent e)
 	{
-		int xPos = paddle.getX();
-		int yPos = paddle.getY();
+		double xPos = paddle.getX();
+		double yPos = paddle.getY();
 	}
 }
